test: 补齐测试覆盖 — 196 → 373 用例 (~25% → 57% 后端覆盖)
新增 6 个测试文件，覆盖之前完全缺失的关键模块：

- test_performance.py (30 用例): P&L、夏普比率、最大回撤、连胜连亏、绩效序列化
- test_repositories.py (33 用例): User/Strategy/Trading 三层 repo 的 CRUD 与定制查询
- test_order_service.py (26 用例): 下单/撤单/止盈止损/平仓/余额同步全生命周期
- test_exchange_adapters.py (44 用例): 适配器工厂、错误分类、Binance 响应解析与状态映射
- test_orders_api.py (26 用例): 订单/账户/紧急平仓 API 鉴权与字段校验
- test_strategies_api.py (18 用例): 策略模板/实例 CRUD/规则校验 API

期间通过测试发现并修复 3 处真实 bug：

1. ExchangeAPIError 未存储 self.exchange 属性 — order_service 在异常分支
   访问 exc.exchange 会抛 AttributeError，导致订单被拒/限流/网络错误时
   无法返回干净的 4xx/5xx，反而 500 崩溃。

2. Binance/OKX/Huobi 三家适配器的 get_ticker / get_klines 都把 Decimal
   传给 datetime.fromtimestamp，触发 TypeError。修复为 float() 转换后再传入。

+++ backend/app/core/exchanges/binance.py
@@ -99,7 +99,7 @@ class BinanceAdapter(BaseExchangeAdapter):
             volume_24h=_safe_decimal(data.get("volume")),
             quote_volume_24h=_safe_decimal(data.get("quoteVolume")),
             timestamp=datetime.fromtimestamp(
-                _safe_decimal(data.get("closeTime"), Decimal("0")) / 1000,
+                float(_safe_decimal(data.get("closeTime"), Decimal("0")) / 1000),
                 tz=timezone.utc,
             ),
         )
@@ -120,13 +120,13 @@ class BinanceAdapter(BaseExchangeAdapter):
         klines = []
         for k in raw:
             klines.append(Kline(
-                timestamp=datetime.fromtimestamp(_safe_decimal(k[0]) / 1000, tz=timezone.utc),
+                timestamp=datetime.fromtimestamp(float(_safe_decimal(k[0]) / 1000), tz=timezone.utc),
                 open=_safe_decimal(k[1]),
                 high=_safe_decimal(k[2]),
                 low=_safe_decimal(k[3]),
                 close=_safe_decimal(k[4]),
                 volume=_safe_decimal(k[5]),
-                close_time=datetime.fromtimestamp(_safe_decimal(k[6]) / 1000, tz=timezone.utc),
+                close_time=datetime.fromtimestamp(float(_safe_decimal(k[6]) / 1000), tz=timezone.utc),
             ))
         return klines
